# swagger_local.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ecommerce Api server for admin and web",
    "version": "1.0.0",
    "title": "Ecommerce"
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/admin/auth/login/admin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login to admin system",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Admin username and password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/member/admin": {
      "get": {
        "tags": [
          "admin member"
        ],
        "summary": "Get all admins",
        "description": "",
        "operationId": "placeOrder",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/member/admin/{adminId}": {
      "get": {
        "tags": [
          "admin member"
        ],
        "summary": "Find admin by id",
        "description": "Returns a single admin",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "put": {
        "tags": [
          "admin member"
        ],
        "summary": "update admin by id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "member update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterAdmin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "delete": {
        "tags": [
          "admin member"
        ],
        "summary": "Deletes an admin",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/admin/order": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "order",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "post an order",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "order update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/order/{orderId}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "order",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "orders"
        ],
        "summary": "update order",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "order update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/order/detail": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "order details",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/order": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "post an order",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "order update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/order/transaction": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "get order transaction",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateOrder": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string",
          "example": "himal"
        },
        "shipAddress": {
          "type": "string",
          "example": "shrestha"
        },
        "shipAddress2": {
          "type": "string",
          "example": "circle"
        },
        "city": {
          "type": "string",
          "example": "circle"
        },
        "state": {
          "type": "string",
          "example": "somwhere wlse"
        },
        "zip": {
          "type": "string",
          "example": "8945"
        },
        "phone": {
          "type": "string",
          "example": "1223234234"
        },
        "country": {
          "type": "string",
          "example": "nepal"
        },
        "tax": {
          "type": "string",
          "example": "far"
        },
        "email": {
          "type": "string",
          "example": "somewhere"
        },
        "paymentId": {
          "type": "integer",
          "format": "int64",
          "example": "1",
          "description": "0 or 1"
        },
        "productId": {
          "type": "integer",
          "format": "int64",
          "example": "1",
          "description": "0 or 1"
        }
      },
      "xml": {
        "name": "UpdateOrder"
      }
    },
...
    "RegisterAdmin": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himalhs1@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "himal123"
        },
        "email": {
          "type": "string",
          "example": "himal123"
        },
        "roles": {
          "type": "string",
          "example": "himal123"
        },
        "fname": {
          "type": "string",
          "example": "himal123"
        },
        "lname": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "RegisterAdmin"
      }
    },
...
    "Auth": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himal"
        },
        "password": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "Pet"
      }
    }
  },
...
}